Finish, but difficult.
One trouble I had was changing the top margins of the .title class to create a gap between it and the .card container.

Adjusting the margins of .title class resulted in a change in gap between the .card container and the top of the page.

margin:"" top-margin:"" all resulted in the .title box spilling over into the .card container.